MAINT: Restructure the project for more granular grouping the sources
