+++ meson.build
@@ -151,49 +151,53 @@ message('  LAPACK found: ' + lapack_msg)
 
 # Build the core library from the translated C sources
 lib_srcs = files(
-  'src/A/ab01nd.c',
-  'src/A/ab04md.c',
-  'src/A/ab05md.c',
-  'src/A/ab05nd.c',
-  'src/A/ab07nd.c',
-  'src/M/ma01ad.c',
-  'src/M/ma01bd.c',
-  'src/M/ma01bz.c',
-  'src/M/ma01cd.c',
-  'src/M/ma01dd.c',
-  'src/M/ma01dz.c',
-  'src/M/ma02ad.c',
-  'src/M/ma02az.c',
-  'src/M/ma02bd.c',
-  'src/M/ma02bz.c',
-  'src/M/ma02cd.c',
-  'src/M/ma02cz.c',
-  'src/M/ma02dd.c',
-  'src/M/ma02ed.c',
-  'src/M/ma02es.c',
-  'src/M/ma02ez.c',
-  'src/M/ma02fd.c',
-  'src/M/ma02gd.c',
-  'src/M/ma02gz.c',
-  'src/M/ma02hd.c',
-  'src/M/ma02hz.c',
-  'src/M/ma02id.c',
-  'src/M/ma02iz.c',
-  'src/M/ma02jd.c',
-  'src/M/ma02jz.c',
-  'src/M/ma02md.c',
-  'src/M/ma02mz.c',
-  'src/M/ma02nz.c',
-  'src/M/ma02od.c',
-  'src/M/ma02oz.c',
-  'src/M/ma02pd.c',
-  'src/M/ma02pz.c',
-  'src/M/ma02rd.c',
-  'src/M/ma02sd.c',
-  'src/M/mb01pd.c',
-  'src/M/mb01qd.c',
-  'src/M/mb03oy.c',
-  'src/M/mc01td.c'
+  # AB family
+  'src/AB/ab01nd.c',
+  'src/AB/ab04md.c',
+  'src/AB/ab05md.c',
+  'src/AB/ab05nd.c',
+  'src/AB/ab07nd.c',
+  # MA family
+  'src/MA/ma01ad.c',
+  'src/MA/ma01bd.c',
+  'src/MA/ma01bz.c',
+  'src/MA/ma01cd.c',
+  'src/MA/ma01dd.c',
+  'src/MA/ma01dz.c',
+  'src/MA/ma02ad.c',
+  'src/MA/ma02az.c',
+  'src/MA/ma02bd.c',
+  'src/MA/ma02bz.c',
+  'src/MA/ma02cd.c',
+  'src/MA/ma02cz.c',
+  'src/MA/ma02dd.c',
+  'src/MA/ma02ed.c',
+  'src/MA/ma02es.c',
+  'src/MA/ma02ez.c',
+  'src/MA/ma02fd.c',
+  'src/MA/ma02gd.c',
+  'src/MA/ma02gz.c',
+  'src/MA/ma02hd.c',
+  'src/MA/ma02hz.c',
+  'src/MA/ma02id.c',
+  'src/MA/ma02iz.c',
+  'src/MA/ma02jd.c',
+  'src/MA/ma02jz.c',
+  'src/MA/ma02md.c',
+  'src/MA/ma02mz.c',
+  'src/MA/ma02nz.c',
+  'src/MA/ma02od.c',
+  'src/MA/ma02oz.c',
+  'src/MA/ma02pd.c',
+  'src/MA/ma02pz.c',
+  'src/MA/ma02rd.c',
+  'src/MA/ma02sd.c',
+  # MB family
+  'src/MB/mb01pd.c',
+  'src/MB/mb01qd.c',
+  'src/MB/mb03oy.c',
+  # MC family
+  'src/MC/mc01td.c'
 )
 
 inc_root = include_directories('.')
