+++ src/AB/ab05nd.c
