+++ src/AB/ab07nd.c
