+++ src/MA/ma01dd.c
