+++ src/MA/ma02dd.c
