+++ src/MA/ma02fd.c
