+++ src/MA/ma02gd.c
