+++ src/MA/ma02od.c
